business-finder/requirements.txt
googlemaps==4.10.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
boto3==1.34.34
selenium==4.18.1
pillow>=10.2.0
python-dotenv==1.0.1
webdriver-manager==4.0.1 